LayoutTests/accessibility/mac/text-markers-for-input-with-placeholder-expected.txt


This tests that start and end text markers of an input element with placeholder are valid.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS input.isEqual(startObject) is true
PASS input.isEqual(endObject) is true
PASS successfullyParsed is true

TEST COMPLETE
